server/requirements.txt
websockets>=12.0
asyncio>=3.4.3
numpy>=1.24.0
pyyaml>=6.0
